requirements.txt
pandas
plotly.express
google.cloud
google.oauth2
streamlit
json
plotly